backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
pix2tex[api]==0.1.2
openai==1.3.0
anthropic==0.7.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-dotenv==1.0.0
numpy==1.26.2
opencv-python==4.8.1.78

